use long int instead

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,16 +10,16 @@
 
 int main(void)
 {
-	long long int x = 0, y = 1, z;
+	long int x = 0, y = 1, z;
 	int n;
 
 	for (n = 0; n < 50; n++)
 	{
 		z = x + y;
 		if (n == 0)
-			printf("%lld", z);
+			printf("%ld", z);
 		else
-			printf(", %lld", z);
+			printf(", %ld", z);
 		x = y;
 		y = z;
 	}
